requirements.txt
# Voice Agent Dependencies

# Core Framework - LiveKit Agents
livekit-agents>=1.0.0
livekit-plugins-groq>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-openai>=1.0.0

# Alternative Framework - Pipecat
pipecat-ai[groq,silero,cartesia,daily]>=0.0.50

# Cloud Providers
groq>=0.11.0
openai>=1.50.0
deepgram-sdk>=3.0.0
cartesia>=1.0.0

# Local Fallbacks
faster-whisper>=1.0.0
ollama>=0.3.0
# kokoro-onnx>=0.4.0  # Install separately: pip install kokoro-onnx
# piper-tts>=1.2.0    # Install separately for local TTS

# Audio Processing
soundfile>=0.12.0
sounddevice>=0.5.0
webrtcvad>=2.0.10
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
httpx>=0.27.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Optional: Telephony
# livekit-plugins-twilio>=1.0.0
